refactor: 의존성 방향 정리 route -> transit
